requirements.txt
fastapi==0.115.2
starlette==0.40.0
uvicorn==0.32.0
anyio==4.6.2.post1
typing_extensions==4.12.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36
databases==0.9.0
googlemaps==4.10.0
openai==1.52.0
requests==2.32.3
python-decouple==3.8
pydantic==2.9.2
pydantic_core==2.23.4
